test: add end-to-end test on cosmian_pkcs11 crate
